Consolidate double dash sentinel test
